[NotoNastaliqUrdu] Fix contour direction of right dot in double-dots
In four glyphs (U+E41B,E415,E41C,E416; gids 16,17,19,20), which are
components used in two/three dots above/below, the two horizontal
dots overlap, and if the contour direction of two dots is different,
the overlap will become hollow.  Fix that.  In general, in TrueType,
all contours are supposed to be clockwise, so this was a double bug.

Test with: ﺕ ﭖ ﺙ ﻱ