Fix up post table in Noto*Lao
All the post tables in Noto*Lao* are broken the same way.

gid=2 through gid=28 have the glyph names of 'nonmarkingreturn' ...
'colon', but they're actually 'space' ... 'semicolon' (off-by-one).
That's because 'nonmarkingreutnr' glyph (gid=2 in the default Mac glyph
order) is missing, but the glyp names are assigned as if it's present.
The following steps were taken (ftxdumperfuser is from Apple's font
tools suite).

  ftxdumperfuser -t post -o lao.post.xml NotoSansLao-Regular.ttf
  # fix up "lao.post.xml"
  # can apply to all other Noto Lao fonts.
  ftxdumperfuser -t post -d lao.post.fixed.xml <Noto Lao font>

[1]
https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6post.html