Drop 'lepc' from GPOS of Noto Sans Lepcha
GPOS for DFLT is identical to that for lepc and there's no need to
keep them both.

GSUB has only 'DFLT' and is left alone.

All the possible syllables were generatd and shaped with and
without 'lepc' in GPOS. No differene was observed for any of them.
(see #451 and #395)

Will fix #451