Update most families with phase 3 versions built from pipeline.
Cham would not autohint, so this removes the hinted Cham so that we
don't have hinted/unhinted versions from different sources.

This increases the number of font files by about 10x.

The new fonts are at 1000 upem, so will not merge with the old fonts
which were at 2048 upem.

NEWS was updated with a bit more information.